Moving the snake to all direction.
Now, the snake can move to left, right, up and down. And also I have changed the shape of snake. If the snake moving rightward and crosses the right border then it will come from the right border. If the snake moving leftward and crosses the left border then it will come from the right border. If the snake moving downward and crosses the border of the down side then it will come from the border of the up side. If the snake moving upward and crosses the border of the up side then it will come from the border of the down side.